requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pymongo==4.6.1
gunicorn==21.2.0
flask-cors==4.0.1
onnxruntime==1.17.1
numpy==1.26.4
requests==2.31.0
transformers==4.40.1
